# api-docs.json
{"openapi":"3.0.1","info":{"title":"Demo API ","description":"Spring Boot Restfull API Demo by VP","termsOfService":"http://localhost:8080/swagger-ui/index.html","contact":{"name":"justmevp","url":"http://localhost:8080/swagger-ui/index.html","email":"vpek215@gmail.com"},"license":{"name":"Apache 2.0","url":"https://www.apache.org/license/LICENSE-2.0"},"version":"Version 1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Album Controller","description":"Controller for Album and Photo management"},{"name":"Auth Controller","description":"Controller for Account management"}],"paths":{"/api/v1/auth/users/{user_id}/update-authorities":{"put":{"tags":["Auth Controller"],"summary":"Update authorities","operationId":"update_auth","parameters":[{"name":"user_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthoritiesDTO"}}},"required":true},"responses":{"200":{"description":"Update authorities","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}},"400":{"description":"Invalid User Id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}},"403":{"description":"Token Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}}},"security":[{"studyeasy-demo-api":[]}]}},"/api/v1/auth/profile/update-password":{"put":{"tags":["Auth Controller"],"summary":"Update profile","operationId":"update_password","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordDTO"}}},"required":true},"responses":{"403":{"description":"Token error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}},"200":{"description":"Update profile","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountViewDTO"}}}}},"security":[{"studyeasy-demo-api":[]}]}},"/api/v1/albums/{album_id}/update":{"put":{"tags":["Album Controller"],"summary":"Update a Album","operationId":"updateAlbum","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumPayloadDTO"}}},"required":true},"responses":{"400":{"description":"Please add a valid name for description","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"204":{"description":"Album added","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"security":[{"studyeasy-demo-api":[]}]}},"/api/v1/albums/{album_id}/photos/{photo_id}/update":{"put":{"tags":["Album Controller"],"summary":"Update a photo","operationId":"update_photo","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoPayloadDTO"}}},"required":true},"responses":{"204":{"description":"Photo added","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoViewDTO"}}}},"400":{"description":"Please add a valid name for description","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoViewDTO"}}}},"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PhotoViewDTO"}}}}},"security":[{"studyeasy-demo-api":[]}]},"delete":{"tags":["Album Controller"],"summary":"Delete a photo","operationId":"delete_photo","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"202":{"description":"Photo Delete","content":{"*/*":{"schema":{"type":"string"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"studyeasy-demo-api":[]}]}},"/api/v1/auth/users/add":{"post":{"tags":["Auth Controller"],"summary":"Add a new User","operationId":"addUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountDTO"}}},"required":true},"responses":{"400":{"description":"Please enter a valid email and Password length between 6 to 20 characters","content":{"application/json":{"schema":{"type":"string"}}}},"200":{"description":"Account added","content":{"application/json":{"schema":{"type":"string"}}}},"201":{"description":"Created","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/api/v1/auth/token":{"post":{"tags":["Auth Controller"],"operationId":"token","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TokenDTO"}}}}}}},"/api/v1/albums/{album_id}/upload-photos":{"post":{"tags":["Album Controller"],"summary":"Upload photo into album","operationId":"photos","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"required":["files"],"type":"object","properties":{"files":{"type":"array","items":{"type":"string","format":"binary"}}}}}}},"responses":{"400":{"description":"Please check the payload or token","content":{"*/*":{"schema":{"type":"array","items":{"type":"object","additionalProperties":{"type":"array","items":{"type":"object"}}}}}}}},"security":[{"studyeasy-demo-api":[]}]}},"/api/v1/albums/add":{"post":{"tags":["Album Controller"],"summary":"Add a new Album","operationId":"addAlbum","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumPayloadDTO"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"200":{"description":"Album added","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"security":[{"studyeasy-demo-api":[]}]}},"/api/v1":{"get":{"tags":["home-controller"],"operationId":"demo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/auth/users":{"get":{"tags":["Auth Controller"],"summary":"List user api","operationId":"users","responses":{"403":{"description":"Token error","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountViewDTO"}}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountViewDTO"}}}}},"200":{"description":"List of users","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountViewDTO"}}}}}},"security":[{"studyeasy-demo-api":[]}]}},"/api/v1/auth/profile":{"get":{"tags":["Auth Controller"],"summary":"View profile","operationId":"profile","responses":{"403":{"description":"Token error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDTO"}}}},"200":{"description":"View profile","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDTO"}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileDTO"}}}}},"security":[{"studyeasy-demo-api":[]}]}},"/api/v1/albums":{"get":{"tags":["Album Controller"],"summary":"List album api","operationId":"albums","responses":{"200":{"description":"List of albums","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"403":{"description":"Token error","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}}},"security":[{"studyeasy-demo-api":[]}]}},"/api/v1/albums/{album_id}":{"get":{"tags":["Album Controller"],"summary":"List album api","operationId":"albums_by_id","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"List of albums","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"403":{"description":"Token error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}},"401":{"description":"Token missing","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AlbumViewDTO"}}}}},"security":[{"studyeasy-demo-api":[]}]}},"/api/v1/albums/{album_id}/photos/{photo_id}/download-thumbnail":{"get":{"tags":["Album Controller"],"operationId":"downloadThumbnail","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"studyeasy-demo-api":[]}]}},"/api/v1/albums/{album_id}/photos/{photo_id}/download-photo":{"get":{"tags":["Album Controller"],"operationId":"downloadPhoto","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"photo_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}},"security":[{"studyeasy-demo-api":[]}]}},"/api/v1/auth/profile/delete":{"delete":{"tags":["Auth Controller"],"summary":"Delete profile","operationId":"delete_profile","responses":{"403":{"description":"Token error","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"Delete profile","content":{"*/*":{"schema":{"type":"string"}}}},"401":{"description":"Token missing","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"studyeasy-demo-api":[]}]}},"/api/v1/albums/{album_id}/delete":{"delete":{"tags":["Album Controller"],"summary":"Delete a album","operationId":"delete_album","parameters":[{"name":"album_id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"202":{"description":"Album Delete","content":{"*/*":{"schema":{"type":"string"}}}},"201":{"description":"Created","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"studyeasy-demo-api":[]}]}}},"components":{"schemas":{"AuthoritiesDTO":{"required":["authorties"],"type":"object","properties":{"authorties":{"type":"string","description":"Authorities","example":"User"}}},"AccountViewDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"authorities":{"type":"string"}}},"PasswordDTO":{"required":["password"],"type":"object","properties":{"password":{"maxLength":20,"minLength":6,"type":"string","description":"Password","example":"Password"}}},"AlbumPayloadDTO":{"required":["description","name"],"type":"object","properties":{"name":{"type":"string","description":"Album name","example":"Travel"},"description":{"type":"string","description":"Description of the album","example":"Description"}}},"AlbumViewDTO":{"required":["description","name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string","description":"Album name","example":"Travel"},"description":{"type":"string","description":"Description of the album","example":"Description"},"photos":{"type":"array","items":{"$ref":"#/components/schemas/PhotoDTO"}}}},"PhotoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"fileName":{"type":"string"},"download_link":{"type":"string"}}},"PhotoPayloadDTO":{"required":["description","name"],"type":"object","properties":{"name":{"type":"string","description":"Photo name","example":"Selfie"},"description":{"type":"string","description":"Description of the photo","example":"Description"}}},"PhotoViewDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"}}},"AccountDTO":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string","description":"Email address","example":"admin@studyeasy.org"},"password":{"maxLength":20,"minLength":6,"type":"string","description":"Password","example":"Password"}}},"UserLoginDTO":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string","description":"Email address","example":"user@user.com"},"password":{"maxLength":20,"minLength":6,"type":"string","description":"Password","example":"pass987"}}},"TokenDTO":{"type":"object","properties":{"token":{"type":"string"}}},"ProfileDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"authorities":{"type":"string"}}}},"securitySchemes":{"studyeasy-demo-api":{"type":"http","in":"header","scheme":"bearer"}}}}